Add back the default theme that has been dropped upstream
We can't remove the theme without advance warnings to users.
Thus we continue to provide it for now and we add a news to tell them
that it will be gone and that they make their own copy of it
if they want to keep it in the long term.