Add a preinst to drop symlinks to directories for tinymce and cropper
The new dh_linktree only symlinks files and hierarchies are duplicated.
So we have to drop symlinks to directories in the preinst, otherwise
dpkg installs the new symlinks in the tinymce/cropper directories
instead of in the wordpress ones.

Also drop the upgrade code in the postinst converting the same directories
into symlinks...

Closes: #639733