Improve dh_linktree
It is now able to generate dependencies and to have different behaviour
for each file to replace. Modify wordpress.linktrees to ensure we have
the very same JQuery files but blindly replaces all the other files.
Drop the explicit dependencies in favor of the autogenerated dependencies.

As a side-effect this fixes installation of widgets which was broken
by the mismatch of some JQuery ui files.