Make the mysql setup script broader in what it accepts.
The script setup-mysql requiered elevated privileges in the mysql server
running on localhost. Both these requirements are not uncommon to be
untrue; the mysql server might be running on another server and the user
can not create the database themselves at the server.
Adding support for optional things but still retaining the old functions
and falling back to them to not break existing documentation surrounding
this package is important.

The added flags enables
- the use of a mysql user not named the same as the database name
  combine -e and -n to achieve this.
- possibility to use a mysql server located on another host.
  use -u and -t